feat: 导航迁移基本完成！！！（大欢呼

+++ src/main/kotlin/cn/spacexc/bilibilisdk/sdk/bangumi/info/remote/Stat.kt
@@ -3,6 +3,6 @@ package cn.spacexc.bilibilisdk.sdk.bangumi.info.remote
 data class Stat(
     val favorites: Int,
     val series_follow: Int,
-    val views: Int,
+    val views: Long,
     val vt: Int
 )
